# openapi.yaml
openapi: 3.0.0
info:
    title: Sistema RC Staff API
    description: API REST frontend-agnostic e desacoplada em Go para o sistema RC Staff (musculação, corrida, RAG, SVED e relatórios administrativos).
    version: 0.1.0-rc
servers:
    - url: http://localhost:5000
      description: Servidor de desenvolvimento local
...
paths:
...
    /api/v1/treinos/mes:
        get:
            operationId: getTreinosMes
            summary: Retorna o calendário mensal de treinos concluídos via link público da Ficha (público)
            tags: [Fichas Web]
            parameters:
                - name: hash_ficha
                  in: query
                  required: true
                  schema:
                    type: string
                    example: a1b2c3d4e5f6
                  description: Hash público da ficha web de musculação. Usado em `/api/v1/treinos/mes`.
                - name: mes
                  in: query
                  schema:
                    type: integer
                    example: 7
                  description: Mês calendário (1–12). Usado em `/api/v1/treinos/mes`.
                - name: ano
                  in: query
                  schema:
                    type: integer
                    example: 2026
                  description: Ano calendário (YYYY). Usado em `/api/v1/treinos/mes`.
            responses:
                '200':
                    description: Frequência mensal do aluno retornada com sucesso
            description: Retorna o calendário mensal de treinos concluídos via link público da Ficha (público) (`GET /api/v1/treinos/mes`).
...
components:
    securitySchemes:
        BearerAuth:
            type: http
            scheme: bearer
            bearerFormat: JWT
            description: 'JWT obtido em POST /api/v1/auth/login. Enviar header Authorization: Bearer <token>.'
...